# api-specs/swagger.json
{
  "info": {
    "title": "Application APIs",
    "version": "1.0.0",
    "description": "Application APIs"
  },
  "basePath": "/",
  "swagger": "2.0",
  "paths": {
...
    "/lesson": {
      "get": {
        "description": "get Lesson",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Lessonresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save Lesson",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "Lesson": {
                  "$ref": "#/definitions/Lesson"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Lessonresponse"
                }
              }
            }
          }
        }
      }
    },
    "/lesson/{id}": {
      "get": {
        "description": "get Lesson with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lesson ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Lessonresponse"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove Lesson with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lesson ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Lessonresponse"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update Lesson with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lesson ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "Lesson": {
                  "$ref": "#/definitions/Lesson"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Lesson",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Lessonresponse"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found",
            "content": {}
          }
        }
      }
    },
    "/person-info": {
      "get": {
        "description": "get PersonInfo",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/PersonInforesponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save PersonInfo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "PersonInfo": {
                  "$ref": "#/definitions/PersonInfo"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/PersonInforesponse"
                }
              }
            }
          }
        }
      }
    },
    "/person-info/{id}": {
      "get": {
        "description": "get PersonInfo with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "PersonInfo ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/PersonInforesponse"
                }
              }
            }
          },
          "404": {
            "description": "PersonInfo not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove PersonInfo with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "PersonInfo ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/PersonInforesponse"
                }
              }
            }
          },
          "404": {
            "description": "PersonInfo not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update PersonInfo with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "PersonInfo ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "PersonInfo": {
                  "$ref": "#/definitions/PersonInfo"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update PersonInfo",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/PersonInforesponse"
                }
              }
            }
          },
          "404": {
            "description": "PersonInfo not found",
            "content": {}
          }
        }
      }
    },
    "/shift": {
      "get": {
        "description": "get Shift",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "get Shift",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Shiftresponse"
                  }
                },
                "pageInfo": {
                  "type": "object",
                  "properties": {
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageLimit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "save Shift",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Shift",
            "schema": {
              "type": "object",
              "properties": {
                "Shift": {
                  "$ref": "#/definitions/Shift"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "save Shift",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Shiftresponse"
                }
              }
            }
          }
        }
      }
    },
    "/shift/{id}": {
      "get": {
        "description": "get Shift with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shift ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "get Shift",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Shiftresponse"
                }
              }
            }
          },
          "404": {
            "description": "Shift not found",
            "content": {}
          }
        }
      },
      "delete": {
        "description": "remove Shift with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shift ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "delete Shift",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Shiftresponse"
                }
              }
            }
          },
          "404": {
            "description": "Shift not found",
            "content": {}
          }
        }
      },
      "put": {
        "description": "update Shift with id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shift ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "Shift",
            "schema": {
              "type": "object",
              "properties": {
                "Shift": {
                  "$ref": "#/definitions/Shift"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "update Shift",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "$ref": "#/definitions/Shiftresponse"
                }
              }
            }
          },
          "404": {
            "description": "Shift not found",
            "content": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Lesson": {
      "properties": {
        "number": {
          "type": "integer"
        },
        "weekDay": {
          "type": "integer"
        },
        "week": {
          "type": "integer"
        },
        "subjectId": {
          "type": "integer"
        },
        "teacherId": {
          "type": "integer"
        },
        "audienceId": {
          "type": "integer"
        },
        "groupId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Lessonresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "number": {
          "type": "integer"
        },
        "weekDay": {
          "type": "integer"
        },
        "week": {
          "type": "integer"
        },
        "subjectId": {
          "type": "integer"
        },
        "teacherId": {
          "type": "integer"
        },
        "audienceId": {
          "type": "integer"
        },
        "groupId": {
          "type": "integer"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "PersonInfo": {
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "middleName": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "PersonInforesponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "middleName": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Shift": {
      "properties": {
        "name": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Shiftresponse": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}